+++ inlang/source-code/paraglide/paraglide-js-adapter-next/examples/app/src/middleware.ts
@@ -1,0 +1,6 @@
+export { middleware } from "@/lib/i18n"
+
+export const config = {
+	//Must be hardcoded string
+	matcher: ["/", `/:path*`],
+}