+++ inlang/source-code/paraglide/paraglide-js-adapter-next/examples/pages/next.config.js
@@ -1,0 +1,13 @@
+const { paraglide } = require("@inlang/paraglide-js-adapter-next/plugin")
+
+/** @type {import('next').NextConfig} */
+module.exports = paraglide({
+	paraglide: {
+		project: "./project.inlang",
+		outdir: "./src/paraglide",
+	},
+	i18n: {
+		locales: ["en", "de"],
+		defaultLocale: "en",
+	},
+})