+++ inlang/source-code/paraglide/paraglide-js-adapter-next/examples/pages/src/pages/_app.tsx
@@ -1,0 +1,10 @@
+import type { AppProps } from "next/app"
+import { ParaglideJS } from "@inlang/paraglide-js-adapter-next/pages"
+
+export default function App({ Component, pageProps }: AppProps) {
+	return (
+		<ParaglideJS>
+			<Component {...pageProps} />
+		</ParaglideJS>
+	)
+}