+++ inlang/source-code/paraglide/paraglide-js-adapter-next/src/app/index.client.tsx
@@ -1,0 +1,1 @@
+export { createI18n } from "./createI18n.client.js"