+++ inlang/source-code/paraglide/paraglide-js/src/cli/commands/init.ts
@@ -95,7 +95,7 @@ export const maybeAddVsCodeExtension = async (args: { projectPath: string }, ctx
 
 	let response = isCertainlyVsCode
 	if (!isCertainlyVsCode) {
-		response = await prompt(`Are you using VSCode?`, {
+		response = await prompt(`Are you using Visual Studio Code?`, {
 			type: "confirm",
 			initial: true,
 		})
@@ -137,7 +137,9 @@ export const maybeAddVsCodeExtension = async (args: { projectPath: string }, ctx
 			"./.vscode/extensions.json",
 			JSON.stringify(extensions, undefined, 2)
 		)
-		ctx.logger.success("Added the inlang vs code extension to the workspace recommendations.")
+		ctx.logger.success(
+			"Added the inlang Visual Studio Code extension (Sherlock) to the workspace recommendations."
+		)
 	}
 }
 
@@ -242,7 +244,7 @@ export const newProjectTemplate: ProjectSettings = {
 		"https://cdn.jsdelivr.net/npm/@inlang/message-lint-rule-valid-js-identifier@latest/dist/index.js",
 		// default to the message format plugin because it supports all features
 		"https://cdn.jsdelivr.net/npm/@inlang/plugin-message-format@latest/dist/index.js",
-		// the m function matcher should be installed by default in case the ide extension is adopted
+		// the m function matcher should be installed by default in case Sherlock (VS Code extension) is adopted
 		"https://cdn.jsdelivr.net/npm/@inlang/plugin-m-function-matcher@latest/dist/index.js",
 	],
 	"plugin.inlang.messageFormat": {