+++ inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @inlang/paraglide-js-adapter-sveltekit
 
+## 0.5.2
+
+### Patch Changes
+
+- @inlang/paraglide-js-adapter-vite@1.2.16
+
+## 0.5.1
+
+### Patch Changes
+
+- a7b593e9a: fix: double-execution of `load` on initial load
+
 ## 0.5.0
 
 ### Minor Changes